refactor: clear 197 CodeQL test-code alerts (Path.Join + redundant-cast bulk transforms) (#172)
* refactor: remove redundant casts in tests (68 CodeQL alerts)

Removes 68 redundant casts flagged by CodeQL across 12 test files:
- 22x cs/useless-cast-to-self (e.g. (uint)expr where expr is already uint)
- 44x cs/useless-upcast on non-null operands (e.g. (T?)rule)
- 2x redundant 'as' upcasts feeding object?-typed reflection helpers

All removed casts are behavior-preserving: they either cast a value to its
own static type, or upcast a non-null operand that flows into a
params object?[] reflection helper where the static type is erased at the
parameter boundary. Method resolution in those helpers is by name + arg
count only, so runtime dispatch is unchanged.

Deliberately NOT touched (left as CodeQL false positives):
- (T?)null casts feeding params object?[] helpers: a bare null collapses to
  the whole params array (TargetParameterCountException), so the cast is
  load-bearing.
- (T?)null inside anonymous-object initializers (CS0828).

Verified: dotnet build (Release) + deterministic test suite green
(3903/3904; the single failure is a pre-existing date/ordering flake in
SupportBundleCoverageSweepTests that passes in isolation).

Part of the QZP drive-to-zero campaign.

* refactor: Path.Combine -> Path.Join in tests (80 CodeQL alerts)

Replaces 80 Path.Combine calls flagged by cs/path-combine
('may silently drop its earlier arguments') with Path.Join across 6 test
files. Audited all 80 sites: every trailing argument is a relative literal
or a locally-constructed relative path; the only absolute component is ever
the FIRST argument (Path.GetTempPath()), where Combine and Join behave
identically. No site has a rooted/absolute second-or-later argument and no
component is empty/whitespace, so the transform is behavior-preserving.

Verified: dotnet build (Release) + full deterministic test suite green
(3904/3904). These are file-system tests that create/read the resulting
paths, so an incorrect path would fail the test - the green run confirms
identical path resolution.

Part of the QZP drive-to-zero campaign.

* refactor: remove redundant (T?)null upcasts in tests (45 CodeQL alerts)

Removes 45 more cs/useless-upcast casts on null literals where null is NOT
the sole argument of a params object?[] reflection helper. With multiple
varargs the compiler wraps them into the array, so the typed null becomes a
normal element[i] and the cast is purely a compile-time annotation that is
erased at the object? parameter boundary - removal is runtime-identical.

Deliberately RETAINED (genuinely load-bearing, NOT removed):
- (T?)null where null is the SOLE params argument: a bare null binds to the
  whole params array, causing TargetParameterCountException. 7 such sites
  identified by the test suite and kept (Invoke/InvokeStatic/InvokeStaticMethod
  with a single null arg).
- (T?)null inside anonymous-object initializers (FlowWave6Tests, CS0828).
These remaining cs/useless-upcast findings are CodeQL false positives at
those specific sites.

Verified: dotnet build (Release) + full deterministic test suite green
(3904/3904).

Part of the QZP drive-to-zero campaign.

---------

+++ tests/SwfocTrainer.Tests/Meg/MegArchiveReaderBranchTests.cs
@@ -558,7 +558,7 @@ public sealed class MegArchiveReaderBranchTests
             writer.Write(0u);
             writer.Write((uint)i);
             writer.Write((uint)entries[i].Bytes.Length);
-            writer.Write((uint)cursor);
+            writer.Write(cursor);
             writer.Write((ushort)i);
             cursor += (uint)entries[i].Bytes.Length;
         }
